create trait SelectorMapElement
Create trait `SelectorMapElement` which will just contain the necessary
functionality for an Element to work with
`SelectorMap::get_all_matching_rules`.

Make `SelectorMapElement` a supertrait of `TElement`. This will ensure
that if there is a function `f<E: TElement>` which calls a function
`g<E: SelectorMapElement>`, and `f` is not needed for using the
SelectorMap, then `f` will not need to be changed to be generic over
`SelectorMapElement`.