restore `traversal_parent` default implementation
Restore default implementation of `traversal_parent`, but move it to
`SelectorMapElement`. This means that `SelectorMapElement` must have an
`as_node` method which returns a `Self::ConcreteNode`. In `TElement`
originally, `Self::ConcreteNode: TNode`, which makes this feature too
specific to `TElement`s. Therefore I think the right approach was to
just bring over `traversal_parent` and remove its default body. Nobody
is going to rely on the default body anyway; we are not rebuilding this
for use in Servo (yet).

This commit just serves as a reference.