add some debugging
Add some debugging code for when I was trying to find problems with
selectors not being matched that should have been in the SelectorMap
algorithm.

One is primarily a print statement to see what is going on, and the
second one is primarily a place to set a breakpoint so I could further
step through the code in a debugger.